controller displays homepage hides the results section on green
