+++ tweetmood/static/js/SpecRunner.html
@@ -6,7 +6,7 @@
 
   <link rel="shortcut icon" type="image/png" href="lib/jasmine-3.2.1/jasmine_favicon.png">
   <link rel="stylesheet" href="lib/jasmine-3.2.1/jasmine.css">
-
+  <script src="https://code.jquery.com/jquery-1.11.3.js"></script>
   <script src="lib/jasmine-3.2.1/jasmine.js"></script>
   <script src="lib/jasmine-3.2.1/jasmine-html.js"></script>
   <script src="lib/jasmine-3.2.1/boot.js"></script>
@@ -16,12 +16,14 @@
   <script src="src/components/form.js" type="text/javascript"></script>
   <script src="src/homepageView.js" type="text/javascript"></script>
   <script src="src/resultsView.js" type="text/javascript"></script>
+  <script src="src/controller.js" type="text/javascript"></script>
 
   <!-- include spec files here... -->
   <script src="spec/buttonSpec.js"></script>
   <script src="spec/formSpec.js"></script>
   <script src="spec/homepageViewSpec.js"></script>
   <script src="spec/resultsViewSpec.js"></script>
+  <script src="spec/controllerSpec.js"></script>
 
 </head>
